requirements.txt
# Основные зависимости
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Веб-скрапинг и обработка данных
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
webdriver-manager>=4.0.0
scrapy>=2.11.0

# Computer Vision и ML
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0  # Для YOLOv11
scikit-learn>=1.3.0
opencv-python>=4.8.0
albumentations>=1.3.1

# Google Gemini и генеративные модели
google-generativeai>=0.3.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0

# API и веб-интерфейсы
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
streamlit>=1.28.0

# Хранение и оркестрация
dagster>=1.5.0
dagster-webserver>=1.5.0
ray>=2.7.0
minio>=7.1.0
wandb>=0.16.0
mlflow>=2.8.0

# Инференс и оптимизация
tritonclient>=2.39.0
optuna>=3.4.0
hyperopt>=0.2.7

# Тестирование и мониторинг
pytest>=7.4.0
evidently>=0.4.5
